+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Forthcoming
 -----------
 * Remove the robot frame property. Instead use the frame from the NavSatFix topic
 * Fix demo.launch
+* Fix uninitialized variable
 
 1.2.0 (2019-03-07)
 ------------------